requirements.txt
streamlit
tensorflow==2.14.0
opencv-python-headless
numpy
Pillow
gdown
